chore: remove hack
Turns out the context pad is updated whenever `elements.changed` is fired so there is no need to update the CSS classes manually. In fact the tests were broken. We have to query the DOM again after `elements.changed` because the entries aren't the same. We also have to use `waitFor` since rendering the the tooltips to wrap the entries is asyncronous.